+++ vizeo_theme/lib/widgets/vz_button.dart
@@ -1,116 +1,125 @@
 import 'package:flutter/material.dart';
+import 'package:vizeo_theme/private/enum.dart';
+import 'package:vizeo_theme/themes/vz_colors.dart';
 import 'package:vizeo_theme/vizeo_theme.dart';
 import 'package:vizeo_theme/private/vz_strings.dart';
 
-enum _TypeButton { PRIMARY, SECONDARY, TERTIARY, SMALL, NOTYPE }
+enum _TypeButton {
+  primary,
+  secondary,
+  tertiary,
+  small,
+  notype,
+}
+
+// TODO
+// mettre une annotation pour dire que si texrtStyleCustom use alors prend le dessu sur le textStyleType
+// aussi pour le text
+// tout sa avant 14H
+// apres bosse projet perso
 
 class VzButton extends StatefulWidget {
-  late final String? myTitle, myFont;
-  late final VoidCallback? onPressedFunc;
-  Color? myBackgroundColor, myHoveredColor, myTextColor;
-  late final double myRadius;
-  late final EdgeInsets myPadding;
-  late final EdgeInsets myIconPadding;
-  final Widget? myRightChild, myLeftChild;
-  late final double myFontSize;
-  late final FontWeight myFontWeight;
+  final String? title;
+  final VoidCallback? onPressed;
+  late final double radius;
+  late final EdgeInsets padding;
+  late final EdgeInsets iconPadding;
   late final bool isEnable;
-  final FocusNode? focus;
   late final _TypeButton type;
+  late final TypeDeTextStyle textStyleBasic;
+  final Widget? rightChild, leftChild;
+  final FocusNode? focus;
+  final TextStyle? textStyleCustom;
+  Color? backgroundColor, hoveredColor, textColor;
 
   VzButton({
-    required this.onPressedFunc,
-    this.myTitle,
-    this.myBackgroundColor,
-    this.myHoveredColor = VzColor.redVizeo,
-    this.myTextColor = VzColor.textPrimaryLight,
-    this.myRadius = 5.0,
-    this.myPadding = const EdgeInsets.all(8.0),
-    this.myIconPadding = const EdgeInsets.all(8.0),
-    this.myRightChild,
-    this.myLeftChild,
-    this.myFontSize = 15,
-    this.myFontWeight = FontWeight.normal,
-    this.myFont = myFontFamily,
+    required this.onPressed,
+    this.title,
+    this.backgroundColor,
+    this.hoveredColor = VzColor.redVizeo,
+    this.textStyleBasic = TypeDeTextStyle.body1,
+    this.padding = const EdgeInsets.all(8.0),
+    this.iconPadding = const EdgeInsets.all(8.0),
+    this.rightChild,
+    this.leftChild,
+    this.radius = 5.0,
     this.isEnable = true,
     this.focus,
-    this.type = _TypeButton.NOTYPE,
+    this.textStyleCustom,
+    this.textColor: VzColor.white,
+    this.type = _TypeButton.notype,
   });
 
   VzButton.typePrimary({
-    required this.onPressedFunc,
-    this.myTitle,
-    this.myBackgroundColor,
-    this.myHoveredColor = VzColor.redVizeoSoft,
-    this.myTextColor = VzColor.textOnRedVizeo,
-    this.myPadding = const EdgeInsets.symmetric(vertical: 10, horizontal: 20),
-    this.myIconPadding = const EdgeInsets.only(left: 12),
-    this.myLeftChild,
-    this.myRightChild,
-    this.myRadius = 5.0,
-    this.myFontSize = 18,
-    this.myFontWeight = FontWeight.w700,
-    this.myFont = myFontFamily,
+    required this.onPressed,
+    this.title,
+    this.backgroundColor,
+    this.hoveredColor = VzColor.redVizeoSoft,
+    this.textStyleBasic = TypeDeTextStyle.body1,
+    this.padding = const EdgeInsets.symmetric(vertical: 10, horizontal: 20),
+    this.iconPadding = const EdgeInsets.only(left: 12),
+    this.leftChild,
+    this.rightChild,
+    this.radius = 5.0,
     this.isEnable = true,
     this.focus,
-    this.type = _TypeButton.PRIMARY,
+    this.textStyleCustom,
+    this.textColor: VzColor.white,
+    this.type = _TypeButton.primary,
   });
 
   VzButton.typeSecondary({
-    required this.onPressedFunc,
-    this.myTitle,
-    this.myBackgroundColor,
-    this.myHoveredColor = VzColor.secondaryButtonSoft,
-    this.myTextColor = VzColor.textOnRedVizeo,
-    this.myPadding = const EdgeInsets.symmetric(vertical: 10, horizontal: 20),
-    this.myIconPadding = const EdgeInsets.only(left: 12),
-    this.myRadius = 5.0,
-    this.myLeftChild,
-    this.myRightChild,
-    this.myFontSize = 18,
-    this.myFontWeight = FontWeight.w700,
-    this.myFont = myFontFamily,
+    required this.onPressed,
+    this.title,
+    this.backgroundColor,
+    this.hoveredColor = VzColor.secondaryButtonSoft,
+    this.textStyleBasic = TypeDeTextStyle.body1,
+    this.padding = const EdgeInsets.symmetric(vertical: 10, horizontal: 20),
+    this.iconPadding = const EdgeInsets.only(left: 12),
+    this.radius = 5.0,
+    this.leftChild,
+    this.rightChild,
     this.isEnable = true,
     this.focus,
-    this.type = _TypeButton.SECONDARY,
+    this.textStyleCustom,
+    this.textColor: VzColor.white,
+    this.type = _TypeButton.secondary,
   });
 
   VzButton.typeTertiary({
-    required this.onPressedFunc,
-    this.myTitle,
-    this.myBackgroundColor,
-    this.myHoveredColor = VzColor.transparent,
-    this.myTextColor = VzColor.textOnRedVizeo,
-    this.myPadding = const EdgeInsets.symmetric(vertical: 10, horizontal: 20),
-    this.myIconPadding = const EdgeInsets.only(left: 8),
-    this.myRadius = 5.0,
-    this.myLeftChild,
-    this.myRightChild,
-    this.myFontSize = 18,
-    this.myFontWeight = FontWeight.w700,
-    this.myFont = myFontFamily,
+    required this.onPressed,
+    this.title,
+    this.backgroundColor,
+    this.hoveredColor = VzColor.transparent,
+    this.textStyleBasic = TypeDeTextStyle.body1,
+    this.padding = const EdgeInsets.symmetric(vertical: 10, horizontal: 20),
+    this.iconPadding = const EdgeInsets.only(left: 8),
+    this.radius = 5.0,
+    this.leftChild,
+    this.rightChild,
     this.isEnable = true,
     this.focus,
-    this.type = _TypeButton.TERTIARY,
+    this.textStyleCustom,
+    this.textColor: VzColor.white,
+    this.type = _TypeButton.tertiary,
   });
 
   VzButton.typeSmall({
-    required this.onPressedFunc,
-    this.myTitle,
-    this.myBackgroundColor,
-    this.myHoveredColor = VzColor.secondaryButtonSoft,
-    this.myTextColor = VzColor.textOnRedVizeo,
-    this.myPadding = const EdgeInsets.symmetric(horizontal: 8),
-    this.myIconPadding = const EdgeInsets.only(left: 12),
-    this.myRadius = 5.0,
-    this.myLeftChild,
-    this.myRightChild,
-    this.myFontSize = 18,
-    this.myFontWeight = FontWeight.w700,
-    this.myFont = myFontFamily,
+    required this.onPressed,
+    this.title,
+    this.backgroundColor,
+    this.hoveredColor = VzColor.secondaryButtonSoft,
+    this.textStyleBasic = TypeDeTextStyle.body1,
+    this.padding = const EdgeInsets.symmetric(horizontal: 8),
+    this.iconPadding = const EdgeInsets.only(left: 12),
+    this.radius = 5.0,
+    this.leftChild,
+    this.rightChild,
     this.isEnable = true,
     this.focus,
-    this.type = _TypeButton.SMALL,
+    this.textStyleCustom,
+    this.textColor: VzColor.white,
+    this.type = _TypeButton.small,
   });
 
   @override
@@ -130,7 +139,6 @@ class _VzButtonState extends State<VzButton> {
   @override
   void initState() {
     super.initState();
-    // change value has focus when user click on TextField
     widget.focus?.addListener(() {
       if (widget.focus!.hasFocus) {
         hasFocus = true;
@@ -141,130 +149,204 @@ class _VzButtonState extends State<VzButton> {
     });
   }
 
-  List<Widget> maListChild() {
+  List<Widget> myListChild() {
     final List<Widget> listChild = [];
-    if (widget.myLeftChild != null) {
-      listChild.add(widget.myLeftChild!);
-      listChild.add(Padding(
-        padding: widget.myIconPadding,
-      ));
+    if (widget.leftChild != null) {
+      listChild.add(
+        widget.leftChild!,
+      );
+      listChild.add(
+        Padding(
+          padding: widget.iconPadding,
+        ),
+      );
     }
-    if (widget.myTitle != null) {
-      listChild.add(Text(
-        widget.myTitle!,
-        style: TextStyle(color: widget.myTextColor),
-      ));
+    if (widget.title != null) {
+      Widget? title;
+      switch (widget.type) {
+        case _TypeButton.primary:
+          title = VzText.typePrimary(
+            data: widget.title!,
+            style: widget.textStyleBasic,
+            textStyleCustom: widget.textStyleCustom,
+            color: widget.textColor,
+          );
+          break;
+        case _TypeButton.secondary:
+          title = VzText.typeSecondary(
+            data: widget.title!,
+            style: widget.textStyleBasic,
+            textStyleCustom: widget.textStyleCustom,
+            color: widget.textColor,
+          );
+          break;
+        case _TypeButton.tertiary:
+          title = VzText.typeTertiary(
+            data: widget.title!,
+            style: widget.textStyleBasic,
+            textStyleCustom: widget.textStyleCustom,
+            color: widget.textColor,
+          );
+          break;
+        case _TypeButton.small:
+          title = VzText.typeSecondary(
+            data: widget.title!,
+            style: widget.textStyleBasic,
+            textStyleCustom: widget.textStyleCustom,
+            color: widget.textColor,
+          );
+          break;
+        case _TypeButton.notype:
+          title = Text(
+            widget.title!,
+            style: const TextStyle(
+              color: VzColor.white,
+            ),
+          );
+          break;
+        default:
+      }
+      listChild.add(title!);
     }
-    if (widget.myRightChild != null) {
-      listChild.add(Padding(
-        padding: widget.myIconPadding,
-      ));
-      listChild.add(widget.myRightChild!);
+    if (widget.rightChild != null) {
+      listChild.add(
+        Padding(
+          padding: widget.iconPadding,
+        ),
+      );
+      listChild.add(
+        widget.rightChild!,
+      );
     }
     return listChild;
   }
 
   ButtonStyle myStyleConfig(BuildContext context) {
     return ButtonStyle(
-      backgroundColor: MaterialStateProperty.resolveWith<Color>((Set<MaterialState> states) {
-        if (widget.type == _TypeButton.PRIMARY) {
-          return widget.isEnable ? widget.myBackgroundColor! : widget.myBackgroundColor!.withOpacity(0.3);
+      backgroundColor:
+          MaterialStateProperty.resolveWith<Color>((Set<MaterialState> states) {
+        if (widget.type == _TypeButton.primary) {
+          return widget.isEnable
+              ? widget.backgroundColor!
+              : widget.backgroundColor!.withOpacity(0.3);
         }
-        if (widget.type == _TypeButton.SECONDARY) {
-          return widget.isEnable ? widget.myBackgroundColor! : widget.myBackgroundColor!.withOpacity(0.3);
+        if (widget.type == _TypeButton.secondary) {
+          return widget.isEnable
+              ? widget.backgroundColor!
+              : widget.backgroundColor!.withOpacity(0.3);
         }
-        if (widget.type == _TypeButton.TERTIARY) {
-          return widget.isEnable ? widget.myBackgroundColor! : widget.myBackgroundColor!.withOpacity(0.3);
+        if (widget.type == _TypeButton.tertiary) {
+          return widget.isEnable
+              ? widget.backgroundColor!
+              : widget.backgroundColor!.withOpacity(0.3);
         }
-        if (widget.type == _TypeButton.SMALL) {
-          return widget.isEnable ? widget.myBackgroundColor! : widget.myBackgroundColor!.withOpacity(0.3);
+        if (widget.type == _TypeButton.small) {
+          return widget.isEnable
+              ? widget.backgroundColor!
+              : widget.backgroundColor!.withOpacity(0.3);
         }
 
-        return widget.isEnable ? widget.myBackgroundColor! : widget.myBackgroundColor!.withOpacity(0.3);
+        return widget.isEnable
+            ? widget.backgroundColor!
+            : widget.backgroundColor!.withOpacity(0.3);
       }),
 
       ///-----Shadow du bouton
-      shape:
-          MaterialStateProperty.all<RoundedRectangleBorder>(RoundedRectangleBorder(borderRadius: BorderRadius.circular(widget.myRadius))),
-      shadowColor: MaterialStateProperty.resolveWith<Color>((Set<MaterialState> states) {
+      shape: MaterialStateProperty.all<RoundedRectangleBorder>(
+        RoundedRectangleBorder(
+          borderRadius: BorderRadius.circular(widget.radius),
+        ),
+      ),
+      shadowColor:
+          MaterialStateProperty.resolveWith<Color>((Set<MaterialState> states) {
         return Colors.transparent;
       }),
 
       side: MaterialStateProperty.resolveWith((states) {
         if (states.contains(MaterialState.focused)) {
-          return widget.isEnable ? BorderSide(color: widget.myBackgroundColor!) : null;
+          return widget.isEnable
+              ? BorderSide(color: widget.backgroundColor!)
+              : null;
         }
       }),
 
       /// style du fond suivant les etats
-      overlayColor: MaterialStateProperty.resolveWith((states) {
-        if (states.contains(MaterialState.pressed)) {
-          if (widget.type == _TypeButton.PRIMARY) {
-            return widget.isEnable ? widget.myBackgroundColor! : widget.myBackgroundColor!.withOpacity(0.3);
+      overlayColor: MaterialStateProperty.resolveWith(
+        (states) {
+          if (states.contains(MaterialState.pressed)) {
+            if (widget.type == _TypeButton.primary) {
+              return widget.isEnable
+                  ? widget.backgroundColor!
+                  : widget.backgroundColor!.withOpacity(0.3);
+            }
+            if (widget.type == _TypeButton.secondary) {
+              return widget.isEnable
+                  ? widget.backgroundColor!
+                  : widget.backgroundColor!.withOpacity(0.3);
+            }
+            if (widget.type == _TypeButton.tertiary) {
+              return widget.isEnable
+                  ? widget.backgroundColor!
+                  : widget.backgroundColor!.withOpacity(0.3);
+            }
+            if (widget.type == _TypeButton.small) {
+              return widget.isEnable
+                  ? widget.backgroundColor!
+                  : widget.backgroundColor!.withOpacity(0.3);
+            }
           }
-          if (widget.type == _TypeButton.SECONDARY) {
-            return widget.isEnable ? widget.myBackgroundColor! : widget.myBackgroundColor!.withOpacity(0.3);
+          if (states.contains(MaterialState.focused)) {
+            return widget.isEnable
+                ? VzColor.white.withOpacity(0.15)
+                : Colors.transparent;
           }
-          if (widget.type == _TypeButton.TERTIARY) {
-            return widget.isEnable ? widget.myBackgroundColor! : widget.myBackgroundColor!.withOpacity(0.3);
+          if (states.contains(MaterialState.hovered)) {
+            return widget.isEnable
+                ? VzColor.white.withOpacity(0.15)
+                : Colors.transparent;
           }
-          if (widget.type == _TypeButton.SMALL) {
-            return widget.isEnable ? widget.myBackgroundColor! : widget.myBackgroundColor!.withOpacity(0.3);
-          }
-        }
-        if (states.contains(MaterialState.focused)) {
-          return widget.isEnable ? VzColor.white.withOpacity(0.15) : Colors.transparent;
-        }
-        if (states.contains(MaterialState.hovered)) {
-          return widget.isEnable ? VzColor.white.withOpacity(0.15) : Colors.transparent;
-        }
-        return Colors.transparent;
-      }),
-
-      /// style du text suivant les etats
-      textStyle: MaterialStateProperty.resolveWith((states) {
-        return TextStyle(
-            color: widget.isEnable ? widget.myTextColor : widget.myTextColor!.withOpacity(0.3),
-            fontSize: widget.myFontSize,
-            fontWeight: widget.myFontWeight,
-            fontFamily: myFontFamily);
-      }),
+          return Colors.transparent;
+        },
+      ),
     );
   }
 
   @override
   Widget build(BuildContext context) {
-    if (widget.type == _TypeButton.PRIMARY) {
-      widget.myBackgroundColor = VzColor.redVizeo;
-    }
-    if (widget.type == _TypeButton.SECONDARY) {
-      widget.myBackgroundColor = VzColor.secondaryButtonColor();
-    }
-    if (widget.type == _TypeButton.TERTIARY) {
-      widget.myBackgroundColor = VzColor.tertiaryButtonColor();
-    }
-    if (widget.type == _TypeButton.SMALL) {
-      widget.myBackgroundColor = VzColor.secondaryButtonColor();
+    switch (widget.type) {
+      case _TypeButton.primary:
+        widget.backgroundColor ??= VzColor.redVizeo;
+        break;
+      case _TypeButton.secondary:
+        widget.backgroundColor ??= VzColor.secondaryButtonColor();
+        break;
+      case _TypeButton.tertiary:
+        widget.backgroundColor ??= VzColor.tertiaryButtonColor();
+        break;
+      case _TypeButton.small:
+        widget.backgroundColor ??= VzColor.secondaryButtonColor();
+        break;
+      default:
     }
     return _button = Container(
       decoration: BoxDecoration(
         borderRadius: BorderRadius.circular(5),
         border: Border.all(
-          color: widget.myBackgroundColor!,
+          color: widget.backgroundColor!,
         ),
-        color: widget.myBackgroundColor,
+        color: widget.backgroundColor,
       ),
       child: ElevatedButton(
-        onPressed: widget.onPressedFunc,
+        onPressed: widget.onPressed,
 
         ///-----Background du bouton
         style: myStyleConfig(context),
         focusNode: widget.focus,
         child: Padding(
-          padding: widget.myPadding,
+          padding: widget.padding,
           child: Row(
             mainAxisSize: MainAxisSize.min,
-            children: maListChild(),
+            children: myListChild(),
           ),
         ),
       ),