+++ vizeo_theme/lib/widgets/vz_checkbox.dart
@@ -4,10 +4,12 @@ import 'package:vizeo_theme/themes/vz_colors.dart';
 class VzCheckbox extends StatefulWidget {
   late bool value;
   late Function(bool) onChanged;
+  bool isEnable;
 
   VzCheckbox({
     required this.value,
     required this.onChanged,
+    this.isEnable = true,
   });
 
   @override
@@ -18,20 +20,23 @@ class _VzCheckboxState extends State<VzCheckbox> {
   @override
   Widget build(BuildContext context) {
     return Checkbox(
-        activeColor: VzColor.backgroundPrimary(isReverse: true),
+        activeColor:
+            widget.isEnable ? VzColor.backgroundPrimary(isReverse: true) : VzColor.backgroundPrimary(isReverse: true).withOpacity(0.4),
         fillColor: MaterialStateProperty.resolveWith((states) {
-          return VzColor.backgroundPrimary(isReverse: true);
+          return widget.isEnable ? VzColor.backgroundPrimary(isReverse: true) : VzColor.backgroundPrimary(isReverse: true).withOpacity(0.4);
         }),
         value: widget.value,
         onChanged: (bool? value) {
-          setState(() {
-            widget.value = !widget.value;
-            if (value != null) {
-              widget.onChanged(value);
-            } else {
-              debugPrint("Heu tu a mal affecter ta variable a value");
-            }
-          });
+          if (widget.isEnable) {
+            setState(() {
+              widget.value = !widget.value;
+              if (value != null) {
+                widget.onChanged(value);
+              } else {
+                debugPrint("Heu tu a mal affecter ta variable a value");
+              }
+            });
+          }
         });
   }
 }