+++ vizeo_theme/CHANGELOG.md
@@ -14,3 +14,31 @@
 ## 0.0.20
 
 * VzTextForm remove width required
+
+## 0.0.43
+
+* bug fix
+* add ENUM for different TextStyle  h1,h2,h3,h4,h5,h6,body1,body2,body3,
+## 0.0.44
+
+* bug fix
+* update VZText & VzTextSelectable for can modifiy TextStyle if user want
+* update VzButton for car update Text inside the button more easily
+
+## 0.0.45
+
+* bug fix
+* correct bug change background color on VzButton
+* add VzFAB button
+
+## 0.0.46
+
+* optimize name
+
+## 0.0.47
+
+* correct borderSize VzFab
+
+## 0.0.48
+
+* correct size VzFab